Modify unit tests for Impute such that Drop is fine.
We fix two issues with the unit tests for Impute to accomodate Drop.
First, all the tuples returned by the MockScanWithMissing had at least
one missing values, so a Drop over all columns would not return any
tuples at all. We fix this by allowing some tuples to be complete
tuples.

Second, the logic in Drop to deal with dropColumns requires the columns
to have names. Thus, we change the TupleDesc created in the mock scan to
use named fields rather than anonymous fields as before.